backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
supabase>=2.5.0
twilio>=9.0.0
httpx>=0.27.0
python-multipart>=0.0.9
apscheduler>=3.10.0
livekit-api>=0.6.0
